libsharedringbuffer: increase size of buffer length to match Rust library
This should have no difference actually since previously the C buff_desc
struct would have had padding anyways with length as a u32. A possibly
better solution is to pack the struct to decrease cache footprint.

+++ echo_server/libsharedringbuffer/include/shared_ringbuffer.h
@@ -20,7 +20,7 @@ typedef void (*notify_fn)(void);
 /* Buffer descriptor */
 typedef struct buff_desc {
     uintptr_t encoded_addr; /* encoded dma addresses */
-    unsigned int len; /* associated memory lengths */
+    uint64_t len; /* associated memory lengths */
     void *cookie; /* index into client side metadata */
 } buff_desc_t;
